requirements.txt
bcrypt==3.1.4
cffi==1.11.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.6.1
python-ldap==3.0.0
six==1.11.0
tornado==5.0
